Move files and update maximum percent diff
